Fruit and Color Recognition
This is a Python program for recognizing fruits and colors entered by the user. The program consists of two functions, 'fruit_reconiser()' and 'color_reconiser()'. The user is prompted to enter a fruit or color name, and the program checks the input against a list of known fruits or colors respectively.